feat(reviewreducer): created reviews reducer
